requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Telegram Bot
aiogram==3.3.0
python-dotenv
pydantic-settings

# Database (New)
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
greenlet  # Нужен для работы SQLAlchemy async